frontend: modify the methods 'submitGuess' and 'changeGuesses'
This commit modifies the 'submitGuess' method so that it does not loop unneccessarily as well as modifying the 'changeGuesses' method so that it does not interfere with the backend. Additionally, this commit modifies the useEffects on the Drawing, Voting, and Results pages to that they include the setState functions used as dependencies. Lastly, this commit adds the shared players array to the Voting page.

* adds the players array to the Voting page (in Room.js)
* modifies the 'submitGuess' and 'changeGuesses' methods in Drawing.js and Voting.js
* modifies the list of dependencies for the 'useEffect' methods in Drawing.js, Voting.js, and Results.js

addresses #36